Move some mixins in the mixinx module
