+++ src/aids/views.py
@@ -25,22 +25,10 @@ from aids.models import Aid, AidWorkflow
 from aids.tasks import log_admins
 from alerts.forms import AlertForm
 from categories.models import Category
-from minisites.mixins import NarrowedFiltersMixin
+from minisites.mixins import SearchMixin, AidEditMixin, NarrowedFiltersMixin
 from programs.models import Program
 from stats.models import Event
 from stats.utils import log_event
-
-
-class SearchMixin:
-    def get_form_kwargs(self):
-        """Take input data from the GET values."""
-
-        kwargs = super().get_form_kwargs()
-        kwargs.update({
-            'data': self.request.GET,
-        })
-
-        return kwargs
 
 
 class AidPaginator(Paginator):
@@ -277,17 +265,6 @@ class AidDetailView(DetailView):
         response = super().get(request, *args, **kwargs)
         log_event('aid', 'viewed', meta=self.object.slug, value=1)
         return response
-
-
-class AidEditMixin:
-    """Common code to aid editing views."""
-
-    def get_queryset(self):
-        qs = Aid.objects \
-            .filter(author=self.request.user) \
-            .order_by('name')
-        self.queryset = qs
-        return super().get_queryset()
 
 
 class AidDraftListView(ContributorRequiredMixin, AidEditMixin, ListView):
